foliounico/requirements.txt
asgiref==3.5.2
psycopg2>=2.8
astroid==2.11.5
backports.zoneinfo==0.2.1
defusedxml==0.7.1
diff-match-patch==20200713
dill==0.3.4
dj-static==0.0.6
Django==4.0.4
django-admincharts==0.4.1
django-admin-list-charts==0.1.3
django-bootstrap5==23.1
django-contrib-comments==2.2.0
django-credit-cards==0.4.1
django-import-export==2.8.0
django-jazzmin==2.5.0
django-phonenumber-field==6.1.0
django-phonenumbers==1.0.1
et-xmlfile==1.1.0
flake8==4.0.1
gunicorn==20.1.0
isort==5.10.1
lazy-object-proxy==1.7.1
MarkupPy==1.14
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.0.10
phonenumbers==8.12.48
platformdirs==2.5.2
Pillow==9.5.0
pycodestyle==2.8.0
pyflakes==2.4.0
pylint==2.13.9
PyYAML==6.0
pytz==2023.3
sqlparse==0.4.2
static3==0.7.0
tablib==3.2.1
tomli==2.0.1
typing_extensions==4.2.0
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0
